dom/media/test/test_seekLies.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: server lies about range requests</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="manifest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onunload="mediaTestCleanup();">
<pre id="test">
<script class="testbody" type="text/javascript">

function on_metadataloaded() {
  var v = document.getElementById('v');
  var d = Math.round(v.duration*1000);
  ok(d == 4000, "Checking duration: " + d);
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
<video id='v'
       preload="metadata"
       src='seekLies.sjs'
       onloadedmetadata='on_metadataloaded();'></video>
</body>
</html>
